Move initializeVersionControl from create.js -> lib/setup.js, fix lint errors
