+++ src/commands/init.js
@@ -7,10 +7,8 @@ import ora from 'ora';
 import chalk from 'chalk';
 import inquirer from 'inquirer';
 import promisify from 'es6-promisify';
-import spawn from 'cross-spawn-promise';
 import path from 'path';
-import which from 'which';
-import { install, initialize, pkgScripts } from './../lib/setup';
+import { install, initialize, pkgScripts, initGit, trimLeft } from './../lib/setup';
 
 const TEMPLATES = {
 	full: 'examples/full',
@@ -24,7 +22,7 @@ export default asyncCommand({
 
 	desc: 'Create a new application interactively',
 
-	async handler(argv) {
+	async handler() {
 		const questions = [
 			{
 				type: 'input',
@@ -81,7 +79,7 @@ export default asyncCommand({
 				message: 'Force option to create the directory for the new app',
 				default: false,
 			}
-		]
+		];
 
 		process.stderr.write('\n');
 		let response = await inquirer.prompt(questions);
@@ -213,7 +211,7 @@ export default asyncCommand({
 		}
 
 		if (response.git) {
-			await initializeVersionControl(target);
+			await initGit(target);
 		}
 
 		return trimLeft(`
@@ -231,58 +229,3 @@ export default asyncCommand({
 		`) + '\n';
 	}
 });
-
-const trimLeft = (string) => string.trim().replace(/^\t+/gm, '');
-
-// Initializes the folder using `git init` and a proper `.gitignore` file
-// if `git` is present in the $PATH.
-async function initializeVersionControl(target) {
-	let git;
-	try {
-		git = await promisify(which)('git');
-	} catch (e) {
-		process.stderr.write('Could not find git in $PATH.\n');
-		process.stdout.write('Continuing without initializing version control...\n');
-	}
-	if (git) {
-		const gitignore = trimLeft(`
-		node_modules
-		/build
-		/*.log
-		`) + '\n';
-		const gitignorePath = path.resolve(target, '.gitignore');
-		await fs.writeFile(gitignorePath, gitignore);
-
-		const cwd = target;
-
-		await spawn('git', ['init'], { cwd });
-		await spawn('git', ['add', '-A'], { cwd });
-
-		const defaultGitEmail = 'developit@users.noreply.github.com';
-		const defaultGitUser = 'Preact CLI';
-		let gitUser;
-		let gitEmail;
-
-		try {
-			gitEmail = (await spawn('git', ['config', 'user.email'])).toString();
-		} catch (e) {
-			gitEmail = defaultGitEmail;
-		}
-
-		try {
-			gitUser = (await spawn('git', ['config', 'user.name'])).toString();
-		} catch (e) {
-			gitUser = defaultGitUser;
-		}
-
-		await spawn('git', ['commit', '-m', 'initial commit from Preact CLI'], {
-			cwd,
-			env: {
-				GIT_COMMITTER_NAME: gitUser,
-				GIT_COMMITTER_EMAIL: gitEmail,
-				GIT_AUTHOR_NAME: defaultGitUser,
-				GIT_AUTHOR_EMAIL: defaultGitEmail
-			}
-		});
-	}
-}
